[IMP] sepa credit tfr: get BIC from bank account if not available on bank
Since the bank is optional on bank accounts and the BIC can be set
on the bank account if there is no bank, we allow the system to
get the BIC from the bank account in case it is not on the bank.

We give priority to the BIC indicated on the bank, keeping
the same behaviour with existing data where the bank BIC is
available.