protect import of external dependencies
Odoo won't install an addon if the external dependencies are not met.
However, the python modules of the addons are imported at startup, and the
lack of an external dependency for an external addon will cause a crash,
therefore the import needs to be in a try..except block.

Conflicts:
	account_banking_sepa_credit_transfer/models/account_banking_sepa.py
	account_banking_sepa_direct_debit/models/banking_export_sdd.py