Mandates multicompany error #242
The restriction rule Banking Mandate multi-company, gives an error when editing a partner bank account from a company A when that account has created a mandate from another company B. So it is impossible to create a mandate for each company without disabling the rule.
Solution: Uncheck the rule writte in Banking Mandate multi-company. But i'm not sure this is correct.