[FIX] account_banking_sepa_credit_transfer: Fix tests due to upstream change
Odoo has added a constraint for avoiding a company currency change if there
are move lines, making these tests to fail, as the currency is changed to EUR.

With this commit, we create a new company with EUR currency for avoiding the
problem. An alternative solution could be to delete all the existing move lines
on the company, but as that can be problematic if there are additional
constraints (extra modules, locking date, etc), I have chosen the other approach.