[FIX] account_banking_sepa_direct_debit: Fix tests due to upstream change
Odoo has added a constraint for avoiding a company currency change if there
are move lines, making these tests to fail, as the currency is changed to EUR.

With this commit, we create a new company with EUR currency for avoiding the
problem.

This commit also changes account_banking_mandate for not duplicating mandate
number, as it was detected during the test creation.

Similar to 1f8e345469695d1fb1a2ba2109ddea3adbdf1f78