[FIX+IMP] account_payment_order: Make tests resilient + inbound inheritable + SavepointCase
Previous tests did several risky operations on tests, like performing an unbound search
on payment orders, what can lead to errors if there are some data on DB (or for example
we add in the future some payment orders in demo).

This commit also split the tests for inbound payment in a base inheritable class that
can be reused in other dependant modules, and the test of this module itself.

Finally, we use SavepointCase for improving performance in tests.