Add sepa and charge_bearer in account_payment_order.
The sepa field is True if all the transactions are sepa payments.

The charge_bearer selection field do nothing. But, when you click on the 'Apply Charge Bearer to all transactions' button, the charge_bearer value is set to all transactions that are not sepa (because sepa require charge_bearer to be SLEV).