[IMP] account_banking_sepa_direct_debit: Remove l10n_generic_coa hidden dependency in tests
As we were setting a new company with its CoA, we need to load in it one, but using
l10n_generic_coa doesn't serve, as we don't depend on that module (and we shouldn't),
but we are invoking its XML-ID for loading the CoA, trusting in the good faith that
the module is available and installed, which is not the case in some cases.

Luckily, Odoo provides already an inheritable test for having a minimal CoA that
doesn't require to load any chart. This PR uses that test and performs the minimal
changes to adapt the module tests to it.