account_payment_order: show payment_reference
The field payment_reference of account.payment is important because it
has the label of the wire transfer. In the 'account' module,
that field is not displayed in tree view nor in form view of
account.payment. This commit fixes this.

account_banking_pain_base: remove empty line introduced by mistake