account_payment_order: re-implement allow_blocked
allow_blocked has been moved from the wizard that adds payment lines to
the payment mode. That way, users cannot modify it. And they cannot
by-pass it by using the button "add to payment order" on the invoice.
The 'blocked' status is also check upon confirmation of the payment
order (in case it has been marked as blocked AFTER being added to
the payment order).