[FIX] account_payment_partner: keep the default value for partner_bank_id
The Odoo default computation for partner_bank_id on customer invoices sets
the first bank account of the company.

The override in account_payment_partner resets it to false if there is no payment mode defined.

Since for customer invoices we usually don't define a payment mode (unless when using direct debit),
this partner_bank_id field is therefore emptied.

This is problematic for instance when generating UBL invoices with the official Odoo module,
because the bank account is mandatory to create a valid Peppol document.