[IMP] account_payment_*: Add support for using payment orders on receipts
When a company uses receipts, for example, to import initial balances (instead of using journal entries), and later needs to pay those receipts, the previous process was complicated.
Before this commit, using a payment order to pay these items was not straightforward.
After this commit, users can generate a payment order for receipts in the same way as for invoices.